requirements.txt
pyyaml>=4.2b1
GitPython==2.1.15
gitdb2==2.0.6
gitdb==0.6.4
Jinja2>=2.10.1
importlib_resources>=1.3; python_version<'3.9'
